backend/requirements.txt
Flask
pypdf
reportlab
pyyaml
cryptography
pytesseract
Pillow